+++ build.config.ts
@@ -1,0 +1,12 @@
+import { defineBuildConfig } from 'unbuild'
+
+export default defineBuildConfig({
+  entries: ['./src/module.ts'],
+  clean: true,
+  declaration: true,
+  rollup: {
+    esbuild: {
+      minify: true,
+    },
+  },
+})