+++ playground/components/content/ProseBlockquote.vue
@@ -1,5 +1,0 @@
-<template>
-  <blockquote class="my-2 border-l-4 border-yellow-200 px-2">
-    <slot />
-  </blockquote>
-</template>