+++ playground/components/content/ProseOl.vue
@@ -1,5 +1,0 @@
-<template>
-  <ol class="mx-2 my-2 list-inside list-decimal">
-    <slot />
-  </ol>
-</template>