+++ playground/components/content/ProseTable.vue
@@ -1,9 +1,0 @@
-<template>
-  <div
-    class="relative mt-10 max-h-96 w-full overflow-x-auto scrollbar:!h-1.5 scrollbar:!w-1.5 scrollbar:bg-transparent scrollbar-track:!bg-themeColor-100 scrollbar-thumb:!bg-themeColor-300"
-  >
-    <table>
-      <slot />
-    </table>
-  </div>
-</template>