+++ playground/components/content/ProseTbody.vue
@@ -1,5 +1,0 @@
-<template>
-  <tbody class="divide-y divide-themeColor-400 border-b">
-    <slot />
-  </tbody>
-</template>