+++ playground/components/content/ProseTd.vue
@@ -1,5 +1,0 @@
-<template>
-  <td class="px-2 text-center">
-    <slot />
-  </td>
-</template>