+++ playground/components/content/ProseTh.vue
@@ -1,5 +1,0 @@
-<template>
-  <th class="">
-    <slot />
-  </th>
-</template>