+++ playground/components/content/ProseUl.vue
@@ -1,5 +1,0 @@
-<template>
-  <ul class="mx-2 my-2 list-inside list-disc">
-    <slot />
-  </ul>
-</template>