+++ playgrounds/content-v3/nuxt.config.ts
@@ -1,0 +1,11 @@
+export default defineNuxtConfig({
+  modules: ['../../src/module', '@nuxt/content'],
+  devtools: { enabled: true },
+  compatibilityDate: '2025-01-01',
+  nuxtToc: {
+    collection: 'content',
+  },
+  future: {
+    compatibilityVersion: 4,
+  },
+})